admin-dashboard/ai/requirements.txt
pinecone-client
chromadb
numpy
pandas
python-dotenv
langchain
langchain-google-genai
google-generativeai
python-dotenv
